Fix for warnings/errors during compilation and linking xdp library on Windows

+++ src/runtime_src/xdp/profile/device/profile_ip_access.h
@@ -21,6 +21,8 @@
 #include <stdexcept>
 #include <fstream>
 #include <iostream>
+#include <string>
+#include "xclhal2.h"
 #include "xclbin.h"
 #include "core/include/xclperf.h"
 #include "xdp_base_device.h"
@@ -53,7 +55,7 @@ public:
      * IP will be requested, otherwise exception will be thrown.
      */
     ProfileIP(Device* handle /** < [in] the xrt or hal device handle */, 
-                int index /** < [in] the index of the IP in debug_ip_layout */, debug_ip_data *data = nullptr);
+              uint64_t index /** < [in] the index of the IP in debug_ip_layout */, debug_ip_data *data = nullptr);
 
     /**
      * The exclusive access should be release in the destructor
@@ -67,14 +69,14 @@ public:
      * (driver) and set the exclusive flag if exclusive access is 
      * granted.
      */
-    virtual void request_exclusive_ip_access(int index);
+    virtual void request_exclusive_ip_access(uint64_t index);
 
     /**
      * The release_exclusive_ip_access API will release the exclusive
      * access granted to this IP to prevent potential card hang, and clear
      * the exclusive flag if success.
      */
-    virtual void release_exclusive_ip_access(int index);
+    virtual void release_exclusive_ip_access(uint64_t index);
 
     /**
      * The map API tries to map the IP specified into user space. The
